% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1.R
\name{L1}
\alias{L1}
\title{L1 splitting function}
\usage{
L1()
}
\value{
Returns a list of functions used in the argument method in rpart
}
\description{
This function provides the L1 splitting function used in rpart. For more details, please refer to the description of the package rpart and its vignettes.
}
\examples{
\dontrun{
library(rpart)
library(rpart.plot)
mystate <- data.frame(state.x77, region=state.region)
names(mystate) <- casefold(names(mystate))
fit1 <- rpart(murder ~ population + illiteracy + income + life.exp +hs.grad + frost + region, data = mystate, method = L1())
rpart.plot(fit1, type =3)
}
}
