% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Max_abs_cor.R
\name{Max_abs_cor}
\alias{Max_abs_cor}
\title{Estimation of Distributions of Maximum Absolute Sample Correlation Coefficient}
\usage{
Max_abs_cor(n = 30, p = 200, replicate = 500)
}
\arguments{
\item{n}{the number of samples}

\item{p}{the dimensions of the standard normal distribution}

\item{replicate}{the number of replication times, i.e., the number of samples for the kernel density estimation}
}
\value{
the same object as the function dkde does
}
\description{
This function provides an estimation of the maximum absolute sample correlation coefficient of standard normal distribution in high dimensions, which is done by generating samples and an Gaussian kernel estimation.
}
\examples{
\dontrun{
set.seed(12345)
result = Max_abs_cor(n = 30, p = 200, replicate = 500)
plot(result)
}
}
