% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Metropolis_C}
\alias{Metropolis_C}
\title{A Metropolis sampler using Rcpp}
\usage{
Metropolis_C(N, sd)
}
\arguments{
\item{N}{the number of samples}

\item{sd}{the standard error of the normal distribution}
}
\value{
a random sample of size \code{n}
}
\description{
A Metropolis sampler using Rcpp used in homework 12
}
