\name{SC19080-package}
\alias{SC19080-package}
\alias{SC19080}
\docType{package}
\title{
  SC19080
}
\description{
  Provide two functions for the final homework of the 'Statistical Computing' course in 2019 fall
}
\details{
  Provide two functions for the final homework of the 'Statistical Computing' course in 2019 fall
}
\author{
Jianrui Zhang.

Maintainer: Jianrui Zhang <zjrustc@mail.ustc.edu.cn>
}
\references{
  See vignetees.
}
\keyword{ package }
\seealso{
  See vignettes.
}

